+++ R/ZenodoRequest.R
@@ -20,6 +20,7 @@ ZenodoRequest <- R6Class("ZenodoRequest",
     requestHeaders = NA,
     data = NA,
     file = NULL,
+    progress = FALSE,
     status = NA,
     response = NA,
     exception = NA,
@@ -55,7 +56,7 @@ ZenodoRequest <- R6Class("ZenodoRequest",
       return(data)
     },
     
-    GET = function(url, request){
+    GET = function(url, request, progress, use_curl = FALSE){
       req <- paste(url, request, sep="/")
       self$INFO(sprintf("Fetching %s", req))
       headers <- c(
@@ -63,19 +64,32 @@ ZenodoRequest <- R6Class("ZenodoRequest",
         "Authorization" = paste("Bearer",private$token)
       )
       
-      r <- NULL
-      if(self$verbose.debug){
-        r <- with_verbose(GET(req, add_headers(headers)))
+      responseContent <- NULL
+      response <- NULL
+      if(use_curl){
+        h <- curl::new_handle()
+        curl::handle_setheaders(h, .list = as.list(headers))
+        response <- curl::curl_fetch_memory(req, handle = h)
+        responseContent = jsonlite::parse_json(rawToChar(response$content))
+        response <- list(
+          request = request, requestHeaders = rawToChar(response$headers), 
+          status = response$status_code, response = responseContent$hits[[1]]
+        )
       }else{
-        r <- GET(req, add_headers(headers))
+        r <- NULL
+        if(self$verbose.debug){
+          r <- with_verbose(GET(req, add_headers(headers), if(progress) httr::progress(type = "up")))
+        }else{
+          r <- GET(req, add_headers(headers), if(progress) httr::progress(type = "up"))
+        }
+        responseContent <- content(r, type = "application/json", encoding = "UTF-8")
+        response <- list(request = request, requestHeaders = headers(r),
+                         status = status_code(r), response = responseContent)
       }
-      responseContent <- content(r, type = "application/json", encoding = "UTF-8")
-      response <- list(request = request, requestHeaders = headers(r),
-                       status = status_code(r), response = responseContent)
       return(response)
     },
     
-    POST = function(url, request, data, file = NULL){
+    POST = function(url, request, data, file = NULL, progress){
       req <- paste(url, request, sep="/")
       if(!is.null(file)){
         contentType <- "multipart/form-data"
@@ -98,14 +112,16 @@ ZenodoRequest <- R6Class("ZenodoRequest",
           url = req,
           add_headers(headers),
           encode = ifelse(is.null(file),"json", "multipart"),
-          body = data
+          body = data,
+          if(progress) httr::progress(type = "up")
         ))
       }else{
         r <- httr::POST(
           url = req,
           add_headers(headers),
           encode = ifelse(is.null(file),"json", "multipart"),
-          body = data
+          body = data,
+          if(progress) httr::progress(type = "up")
         )
       }
       
@@ -115,7 +131,7 @@ ZenodoRequest <- R6Class("ZenodoRequest",
       return(response)
     },
     
-    PUT = function(url, request, data){
+    PUT = function(url, request, data, progress){
       req <- paste(url, request, sep="/")
       
       if(regexpr("api/files", req)<0) data <- private$prepareData(data)
@@ -132,13 +148,15 @@ ZenodoRequest <- R6Class("ZenodoRequest",
         r <- with_verbose(httr::PUT(
           url = req,
           add_headers(headers),    
-          body = data
+          body = data,
+          if(progress) httr::progress(type = "up")
         ))
       }else{
         r <- httr::PUT(
           url = req,
           add_headers(headers),    
-          body = data
+          body = data,
+          if(progress) httr::progress(type = "up")
         )
       }
       
@@ -181,10 +199,11 @@ ZenodoRequest <- R6Class("ZenodoRequest",
     #' @param request the method request
     #' @param data payload (optional)
     #' @param file to be uploaded (optional)
+    #' @param progress whether a progress status has to be displayed for download/upload
     #' @param token user token
     #' @param logger the logger type
     #' @param ... any other arg
-    initialize = function(url, type, request, data = NULL, file = NULL,
+    initialize = function(url, type, request, data = NULL, file = NULL, progress = FALSE,
                           token, logger = NULL, ...) {
       super$initialize(logger = logger)
       private$url = url
@@ -192,6 +211,7 @@ ZenodoRequest <- R6Class("ZenodoRequest",
       private$request = request
       private$data = data
       private$file = file
+      private$progress = progress
       private$token = token
      
     },
@@ -200,9 +220,10 @@ ZenodoRequest <- R6Class("ZenodoRequest",
     execute = function(){
       
       req <- switch(private$type,
-        "GET" = private$GET(private$url, private$request),
-        "POST" = private$POST(private$url, private$request, private$data, private$file),
-        "PUT" = private$PUT(private$url, private$request, private$data),
+        "GET" = private$GET(private$url, private$request, private$progress),
+        "GET_WITH_CURL" = private$GET(private$url, private$request, private$progress, use_curl = TRUE),
+        "POST" = private$POST(private$url, private$request, private$data, private$file, private$progress),
+        "PUT" = private$PUT(private$url, private$request, private$data, private$progress),
         "DELETE" = private$DELETE(private$url, private$request, private$data)
       )